+++ frontend/src/store/state.js
@@ -1,11 +1,6 @@
 import { reactive } from 'vue';
 import { detectLocale } from "@/i18n";
 
-function loadDisplayPreferences() {
-  const history = localStorage.getItem("displayPreferences");
-  return history ? JSON.parse(history) : {};
-}
-
 export const state = reactive({
   tooltip: {
     show: false,
@@ -29,7 +24,7 @@ export const state = reactive({
   isMobile: window.innerWidth <= 800,
   isSearchActive: false,
   showSidebar: false,
-  displayPreferences: loadDisplayPreferences(),
+  displayPreferences: {},
   usages: {},
   editor: null,
   serverHasMultipleSources: false,