+++ frontend/vite.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import VueI18nPlugin from "@intlify/unplugin-vue-i18n/vite";
 import { compression } from "vite-plugin-compression2";
+import checker from "vite-plugin-checker";
 
 const plugins = [
   vue(),
@@ -12,6 +13,17 @@ const plugins = [
   compression({
     include: /\.(js|woff2|woff)(\?.*)?$/i,
     deleteOriginalAssets: true,
+  }),
+  checker({
+    typescript: {
+      buildMode: true,
+    },
+    vueTsc: {
+      tsconfigPath: "./tsconfig.json",
+    },
+    eslint: {
+      lintCommand: 'eslint "./src/**/*.{js,vue,ts}"',
+    },
   }),
 ];
 
@@ -46,7 +58,7 @@ export default defineConfig(({ command }) => {
         if (hostType === "js") {
           return { runtime: `window.__prependStaticUrl("${filename}")` };
         } else if (hostType === "html") {
-          return `{{ .staticURL }}/${filename}`;
+          return `{{ .htmlVars.staticURL }}/${filename}`;
         } else {
           return { relative: true };
         }